refactor: add sync class and clean up extension file. it should improve startup performance #472
